refactor: better handle missing config, and add test
