+++ BALSAMIC/utils/cli.py
@@ -292,12 +292,10 @@ def get_config(config_name):
 
     p = Path(__file__).parents[1]
     config_file = str(Path(p, 'config', config_name + ".json"))
-    try:
-        Path(config_file).exists()
+    if Path(config_file).exists():
         return config_file
-    except FileNotFoundError as err:
-        raise err
-
+    else:
+        raise FileNotFoundError(f'Config for {config_name} was not found.')
 
 
 def get_ref_path(input_json):
